Restrict the branch copy heuristic
The FreeBSD project has a different approach to branches than the
standard SVN or GIT models of how branches should work. An MFC from
head/ to stable/8 is a cherry-pick in git and should never result in a
merge commit.

We still would want to track merges between user/, project/ and master
branches, however, so that "IFC"s and such are recorded correctly.

Reported by: Ryan Stone <rysto32@gmail.com>